Deconvolved cell-type proportions from STdeconvolve for a Visium data from the adult mouse brain along with the positions of their respective spots

Deconvolved cell-type proportions from STdeconvolve for a Visium data from the adult mouse brain along with the positions of their respective spots

Usage

data(adult_mouse_brain_ffpe)

Format

A list of 2 objects, each of 2264 rows and 2 columns: prop, which contains the 12 deconvolved cell-types proportions for each 2264 spots/pixels in a spatial transcriptomics experiment performed on adult mouse brain tissue tissue and pos, which contains the x and y-coordinates for each 2264 spots/pixels on an adult mouse brain tissue slide.

Source

https://www.10xgenomics.com/datasets/adult-mouse-brain-ffpe-1-standard-1-3-0